+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 0.4.1
     + Separate managers from models
+    + Add search_data field and update it on save
 
 0.4.0
     + Add langauge param to Post's get_absolute_url method